refactor:PuppyDogAnimationの歩きモーション分離、型定義
