+++ components/animations/PuppyDogAnimation.tsx
@@ -14,6 +14,8 @@ import earImage from '../../public/ダックス耳.png';
 import earImageRight from '../../public/ダックス奥耳.png';
 import jawImage from '../../public/ダックス顎.png';
 
+import { usePuppyDogWalkingAnimation } from '../../hooks/components/animations/puppyDog/usePuppyDogWalkingAnimation'
+
 const PuppyDogAnimation = () => {
   const legBackLeftRef = useRef<HTMLImageElement | null>(null);
   const legBackRightRef = useRef<HTMLImageElement | null>(null);
@@ -37,74 +39,6 @@ const PuppyDogAnimation = () => {
   const getRandomPosition = () => {
     const viewportWidth = window.innerWidth;
     return Math.random() * -(viewportWidth - viewportWidth / 2);
-  };
-
-  const startWalkingAnimation = () => {
-    const repeat = -1;
-    const yoyo = true;
-
-    const legAnims = [
-      gsap.to(legBackLeftRef.current, {
-        rotation: 30,
-        transformOrigin: 'top',
-        duration: 1.0,
-        repeat,
-        yoyo,
-        ease: 'power1.inOut',
-        delay: 0, //待機時間
-      }),
-      gsap.to(legBackRightRef.current, {
-        rotation: -30,
-        transformOrigin: 'top',
-        duration: 1.0,
-        repeat,
-        yoyo,
-        ease: 'power1.inOut',
-        delay: 0.6,
-      }),
-      gsap.to(legFrontLeftRef.current, {
-        rotation: 20,
-        transformOrigin: 'top',
-        duration: 1.0,
-        repeat,
-        yoyo,
-        ease: 'power1.inOut',
-        delay: 1.2,
-      }),
-      gsap.to(legFrontRightRef.current, {
-        rotation: -20,
-        transformOrigin: 'right',
-        duration: 1.0,
-        repeat,
-        yoyo,
-        ease: 'power1.inOut',
-        delay: 1.8,
-      }),
-    ];
-
-    const tailAnim = gsap.to(tailRef.current, {
-      rotation: 5,
-      transformOrigin: 'bottom',
-      duration: 0.5,
-      repeat: -1,
-      yoyo: true,
-    });
-
-    const headAnim = gsap.to([headFaceRef.current, headEyeRef.current, bodyRef.current, earRef.current, earRightRef.current, jawRef.current], {
-      y: 5,
-      duration: 1,
-      repeat: -1,
-      yoyo: true,
-    });
-
-    // 初回のアニメーションを開始
-    animate();
-
-    return () => {
-      legAnims.forEach(anim => anim.kill());
-      tailAnim.kill();
-      headAnim.kill();
-    };
   };
 
   const animate = () => {
@@ -133,6 +67,21 @@ const PuppyDogAnimation = () => {
       },
     });
   };
+
+  const { startWalkingAnimation } = usePuppyDogWalkingAnimation({
+    legBackLeftRef,
+    legBackRightRef,
+    legFrontLeftRef,
+    legFrontRightRef,
+    tailRef,
+    headFaceRef, 
+    headEyeRef, 
+    bodyRef, 
+    earRef, 
+    earRightRef, 
+    jawRef,
+    animate
+  });
 
   useEffect(() => {
     if (containerRef.current) {
